added user manager to the new site system
