+++ config/config.php
@@ -520,7 +520,8 @@ $ne_menu = array(
 $ne_site_name = array(
     "not_found"       => "Seite nicht gefunden!",
     "areas_manager"   => "Bereich Management",
-    "nav_editor"      => "Seite und Navigation"
+    "nav_editor"      => "Seite und Navigation",
+    "user_manager"    => "User management"
 );
